src/backend/requirements.txt
fastapi
uvicorn
chromadb
openai
pydantic
python-dotenv